chore(query): combine and document
