+++ packages/core/src/components/query/readme.md
@@ -11,30 +11,39 @@
 
 string
 
+Bind the elements visibility to any valid CSS media query.
+This uses `matchMedia` under the hood.
+
 
 #### mode
 
 string
+
+When the app's mode matches the value, the element will show. This differs
+from `platform` as you could be on iOS, but use the `md` mode.
 
 
 #### orientation
 
 string
 
+When the device's orientation matches the value, the element will show
+
 
 #### platform
 
 string
 
-
-#### queryMatches
-
-boolean
+When the device OS matches the value, the element will show. Examples are
+"ios" or "android"
 
 
 #### size
 
-string
+
+
+Shorthand method for showing an element based on preset sizes.
+Valid sizes are `xs`, `sm`, `md`, `lg`, or `xl`.
 
 
 ## Attributes
@@ -43,30 +52,46 @@ string
 
 string
 
+Bind the elements visibility to any valid CSS media query.
+This uses `matchMedia` under the hood.
+
 
 #### mode
 
 string
+
+When the app's mode matches the value, the element will show. This differs
+from `platform` as you could be on iOS, but use the `md` mode.
 
 
 #### orientation
 
 string
 
+When the device's orientation matches the value, the element will show
+
 
 #### platform
 
 string
 
-
-#### query-matches
-
-boolean
+When the device OS matches the value, the element will show. Examples are
+"ios" or "android"
 
 
 #### size
 
-string
+
+
+Shorthand method for showing an element based on preset sizes.
+Valid sizes are `xs`, `sm`, `md`, `lg`, or `xl`.
+
+
+## Methods
+
+#### doesMatch()
+
+Method to check if the element currently matches the specific queries.
