+++ packages/core/src/components/query/test/basic/index.html
@@ -69,23 +69,23 @@
 
         <h2>Size Tests</h2>
         <ion-query size="xs">
-          <div>Only render on xs</div>
+          <div>Renders on xs</div>
         </ion-query>
 
         <ion-query size="sm">
-          <div>Only render on sm</div>
+          <div>Renders on sm</div>
         </ion-query>
 
         <ion-query size="md">
-          <div>Only render on md</div>
+          <div>Renders on md</div>
         </ion-query>
 
         <ion-query size="lg">
-          <div>Only render on lg</div>
+          <div>Renders on lg</div>
         </ion-query>
 
         <ion-query size="xl">
-          <div>Only render on xl</div>
+          <div>Renders on xl</div>
         </ion-query>
 
         <h2>Media Query</h2>
